Make aggregation queries return zero-value rows
Specifically, if you query an aggregation and give all the group-by
values, but it doesn't match anything, a single, zero-value result will
be given. Unfortunately, this isn't enough on its own. If the node is
materialized (which aggregations should be), then query() will never be
used, they just hit the materialized view. Thus, no zero-value will be
returned. Zero-value rows are also not emitted for a "query all" style
query, because the group-by identifiers for zero-value rows aren't
known. This means that if the aggregation is used as the outer loop in a
join, no zero-value results will be produced. This is technically the
correct operation for a JOIN, but we probably really want the behavior
of a LEFT JOIN in this case.