Remove the ts argument from end-user queries
Ultimately, this was done so that backlog can avoid scanning the buffer
and only do lookups into the Store when a user query is being serviced.
This speeds things up, because the index-lookup in Store will be faster
if the backlog does not also need to be consulted. This is especially
true if the backlog contains negative facts, as every positive fact
must be compared about all the negative facts before being emitted.

Unfortunately, this change has the side-effect that user queries now see
stale data, even if they wait after writing. In particular, a write
won't be visible to queries until a) all descendants of the view being
queried have seen a *later* write, and b) the view being queried
realizes that all its descendant views' mins have progressed.

This has two implications. First, it fixes the behavior in the previous
version where a query could potentially see inconsistent results.
Specifically, if multiple updates with the same timestamp ended up in
the graph (which could happen without there being bugs), queries could
observe state with one applied, but not the other. Now, a write is
visible only once a write with a later timestamp has been received.

For inner queries (i.e., end-user queries that target non-leaf views),
this introduces a weird kind of read-your-own-writes. Specifically, a
query will observe a write only once the view being queried has
*realized* that all its children have seen a *later* write. Logically,
this occurs because a node can't absorb state beyond the earliest time
it may still receive queries for. Thus, if a write w1 arrives at some
node n1 with ts1, then it can't be absorbed until n1 knows that none
of its descendants (say, n2) will query at ts < ts1. This will be true
once n2 receives a write w2 with some ts2 > ts1, but n1 might not learn
about it immediately. In particular, nodes only check whether the
descendant timestamp has progressed immediately after processing an
update. Thus, if n1 doesn't see that n2 is at ts2 *while* n1 is
processing w2, a *second* write, w3, is needed for w1 to observe n2's
timestamp change, and thus to absorb w1 (which again exposes it to
end-user queries).

It might be obvious, but this change in semantics breaks a bunch of the
tests, because they assume that after they've written a bunch of things,
they can see the results by waiting for a while and then querying. This
is no longer true. Test fixes are in subsequent commits.