Add workarounds for broken non-mat query behavior
The query that removed timestamps from end-user queries also exposed
another issue. In particular, all end-user queries that hit a
non-materialized view are given a timestamp of `i64::max_value()`, which
means they *always* see all the data. This conflicts with the behavior
seen when querying a materialized view. However, more importantly, it
can yield inconsistent results. For example, if a join is queried in
this way, the left and right sides of the join may return results from
different timestamps, which could yield inconsistencies.

Unfortunately, it is not entirely clear what the fix for this is. When a
non-materialized view receives an end-user query, what timestamp is
correct to assign to it? In the current system, *any* timestamp that
isn't effectively i64::max_value() may already have been absorbed by the
time it reaches the ancestor view being queried, which will crash the
system. We could make it not crash the system, but that would yield
inconsistencies. We effectively need a mechanism for ensuring that the
system doesn't absorb while we're querying, but that's not entirely
straightforward to add. Brainstorming needed.

This commit adds workarounds to the appropriate tests such that they
check for this weird alternative behavior. When this is fixed, the
statements inside if false should replace the entire containing if.