Stop indexing aggregation output column
self.over refers to the *input* field, not the *output* field of the
aggregation. This caused issues, because the query used to look up in
the aggregation output (currently) includes a None argument for the
output field. Shortcut would then decide to use the index on the output
field, which is no longer a unique lookup!