Rudimentary support for printing flow graphs.
This adds code that implements `Debug` for the FlowGraph and related
structures. The flow graphs are printed in GraphViz/Dot syntax for
visualization. This is helpful for debugging on larger and transformed
graphs. Unfortunately, we can't easily implement `Display` because we
store the graph nodes and edges as `Option<Arc<View>>` and
`Option<Arc<Query>>` types, and `Option` does not implement `Display`.
As a result, the printed flow graphs are quite verbose. We could use a
custom wrapper type, or customize the `Debug` output for our operators
to fix this.