Write basic web frontend
Try it with:

    curl --data-raw '{"id": 1, "title": "hello"}' -H "Content-Type: application/json" localhost:8080/article
    curl --data-raw '{"user": 1, "id": 1}' -H "Content-Type: application/json" localhost:8080/vote
    curl localhost:8080/awvc

Keep in mind that writes aren't seen until at least one write has
bypassed them. We may want to fix this in the future, but for now, just
do a few more writes and you'll see read results that lag a little
behind.

The biggest shortcoming of the current draft is that it doesn't provide
query parameters. It simply queries for *all* results. This will be
fixed.