Unfinished work on condition trees.
This change makes it possible to, in principle, parse arbitrary
condition trees of logical operators and comparison operators. However,
there is a bug that causes the `many0!()` invocation inside the
`boolean_primary` rule to attempt to parse the terminator (semicolon or
newline) of a query in which the WHERE clause is the final part of the
query (as, e.g., in the `where_clause` test case).