Enable nodes to buffer within a timestamp
This commit changes the way `process()` works, so that a node can
indicate that refuse to release an update for a timestamp until it has
been sent more updates within that same timestamp.

This became necessary as a result of
99bf70ab413190cbdb81ad28f13c05d044610936, which enforced the property
that a node may only emit a single update per timestamp. This broke
Union, which depends on being able to emit all the updates it receives
from its ancestors for a given timestamp. Following
99bf70ab413190cbdb81ad28f13c05d044610936, Union would need to *query*
its ancestors for all its updates *at a given timestamp* when it
received the first update for a timestamp, and then throw away the same
updates when it receives them a short while later. This is inefficient.

With this patch, a node (such as Union), can declare that it is not yet
ready to produce an update for a timestamp. If it does, the node will
not send anything to its children, but instead continue consuming
updates from its parents for the same timestamp. Only when the node
indicates, for some later update, that it is willing to commit to an
update for the timestamp (i.e., when it has heard from all the ancestors
for that timestamp) will the node forward the resulting update.