Working nested conditional expressions
We can now build condition trees from WHERE clauses with (somewhat)
arbitrary nested conditions.

The key insight was that the `chain!()` parsers in the conditional
expressions need to use `complete!()` to fail with an error if they
cannot parse any further (e.g., because they encounter a teminator or
the next clause).