Work translating conditions to filter nodes
This is checkpointing unfinished work on parsing conditionals with
literal values into filter nodes (identity operators with output
filters) in the data-flow graph. It doesn't currently validate that a
conditional only contains literals, and also does not automatically
split conditionals.

Checkpointing this for the moment in order to do the more substantive
re-write of FromSql into ToFlowParts.