ToFlowParts trait to add SQL query into flow graph
This refactors the SQL query integration code to use the ToFlowParts
trait (here, implemented on &str) to convert from a query string to a
set of nodes added to a flow graph. This is a better design because it
avoids having helper methods as part of the trait, and because the trait
applies to the type being converted.

This commit only refactors; functionality is identical to the prior
commits.