Fix incorrect timestamps on delayed updates
This fixes a bug introduced in d145479006d1601daae755810ffc21503eff7a24,
which manifested itself as a non-deterministic crash in the "vote"
benchmark. The problem was that delayed updates at a node would *all* be
processed at the current (later) incoming update's timestamp, rather
than their original timestamps. This led to `process()` being called
with the wrong timestamp, which caused early absorption to the (later)
current incoming update timestamp, and a subsequent crash when trying to
add more delayed items to the node's backlog at older timestamps.

In terms of the code, this happened because the `term()` closure
captured `ts` from its environment, rather than using the `ts` value at
its call sites. By passing the correct `ts` as an explicit argument, we
get the right behaviour. This fixes the "vote" benchmark, which used to
crash in the prepopulation phase.