fix up tests/lib.rs for domains branch
note that we don't actually have a way to do gets yet, so these tests
all fail with a 'key not found' error on the 'get' map returned by
commit().

to fix this we'll need to add support *shared* materialized edge view
state and/or support for queries reaching into a domain (if the view the
query is for isn't materialized). it's unclear how best to do this.