remove query() entirely
we have no graphs that actually use the ability to internally query
non-materialized nodes. arguably, they should simply be materialized
instead. this gets rid of a bunch of code and tests that we basically
never exercised, as well as a number of code paths that used to call
unreachable!() when they would otherwise have queried another node