internal materialized tables are now just hashmaps
some decent speedups
this does place an additional restriction on nodes in that they can only
do lookups in their state on a single key. we don't have graphs that use
multiple keys, but we had the code and tests to support it. this patch
removes those, with the hope that we eventually add it back in.