Arc records to avoid cloning in some places
We might want to use Rc inside thread domains and only Arc between them
eventually, but it's unclear it'll matter. We also still need to clone
in some places. In particular:

 - Joins need to create new output records
 - Grouped operators need to construct new records for Positives
 - Permute and union need to create new records when then subset a rec

We save in the following ways though:

 - Aggregation negatives are now not cloned (they just re-use the record
   that was in the materialized state).
 - Latest re-uses for both negative (the materialized) and positive (the
   incoming)
 - Output that is materialized and that is also sent to children is not
   cloned for insertion into the materialized state.

This change is also useful in that it forces the various operators to
assume that their input is immutable. This will make it easier to move
the code either to Rc, or to some other memory-saving technique later.