Recover if a commit object no longer exists
With forced pushes that rewrite history, it's possible that a commit
object no longer exists at the point when taster attempts to move the
workspace to it. There's not much we can do in this situation, other
than emit an error message and continue service; that's what we do now
(previously, taster would simply crash).