Force chain! invocations to complete.
This fixes the unit tests that combine a WHERE clause with ORDER BY or
LIMIT clauses: previously, a parser could eagerly recurse into a rule by
matching the space at its beginning and then give up with an Incomplete
error if it hit the end of the query unexpectedly, which would terminate
the entire parsing. complete! forces the parser to return an error in
this case, so that subsequent parsers are tried as well.

Also correctly makes leading spaces of clauses optional and moves some
of them around (it seems it's better to have space? inside a following
opt! rule).

All tests pass again!