Query signatures based on query graph structure
This adds a `QuerySignature` struct that we will use to reject
unsuitable temporaries according to the Finkelstein algorithm. The query
signature contains both sets of relations and attributes (= columns)
contained in a query graph and a hash of the QG structure. The latter
can be used to detect entirely identical query graphs; the former are
useful for subset checks.