Very simple pass to expand stars in SQL queries.
This expands any instance of `FieldExpression::All` to the corresponding
explicit list of column names. This requires us to maintain a map of
write schemas based on the base nodes we have extracted from INSERT
queries, as otherwise there is no authoritative way of knowing which
columns exist in a write of a specific type. (We might infer them from
the set of known queries in the future, but in practice, there will
always be an explicit write/INSERT specified for any base table that
other views rely upon, at least currently.)