Stateful SqlIncorporator to enable rewrite passes
This is required because the `StarExpansion` and
`ImpliedTableExpansion` passes require access to the column schemas of
the writes (INSERT queries) that exist in the Soup already. SQL queries
now have to be added via a `SqlIncorporator`, who stores a mutable
reference to the `FlowGraph` (i.e., no manual node incorporation is
possible as long as the `SqlIncorporator` holds the `FlowGraph` ref).
Note that we only know the schemas for writes that were added as SQL
queries, not manually added ones, so an application can't mix the two.