Implied table interference pass for SQL queries.
This adds a pass that expands "implied" (as in, not explicitly named)
tables in SQL queries: these can occur in the list of output columns (as
in "SELECT name FROM users") or in conditions (as in "WHERE author =
uid" for "WHERE articles.author = users.uid"). Such short-hand is common
in the TPC-W queries.

In this pass, we expand all mentions of columns such that their `Column`
structure contains an explicit `table`.