Workaround for non-determinstic join reordering
This provisionally fixes a non-deterministic failure when building
graphs from a query with multiple joins: depending on the order in which
the joins are added, and depending on the order of columns in the
condition trees that describe the join conditions, we could end up
looking for a field in the wrong column (on the "wrong side" of the
join). I *think* sorting both the condition tree arms and the joins into
a known order before processing *should* fix this, but I'll need to give
it a bit more thought.

We now successfully incorporate 28 queries out of TPC-W into a Soup
graph.