Expose leaf view per SQL query, add named views
This adds a final missing component to building graphs from SQL
queries: the generation of the leaf/edge view that actually represents
the entire query and its results. This is where readers ought to read
the results of this query; all other nodes are anoymous internal nodes.

SQL SELECT queries have no natural identifier, so we return a generated
one by default now. However, generated identifiers are painful when
constructing graphs from SQL queries, so we also support passing an
optimal explicit view name with `to_flow_parts`. This name is used for
the leaf node of the query only.