Add first draft of cross-domain replay
Some serious limitations:

 - Doesn't replay materialized state for Readers
 - Doesn't handle replay for nodes with multiple parents.

The latter is a kind of interesting case, because the right thing to do
depends on the type of the node. For a Union, you need to replay all
paths. For a Join, you should only replay one side. Unclear what the
pattern is, or if we should special-case all operators?