Avoid exposing partial migration replays
This patch modifies the migration process so that Reader state is only
swapped (and thus exposed) once all state has been replayed. This in
turn enables Reader nodes to return Err() when they do not yet have any
state (as opposed to returning an empty Vec for rows that have not yet
been populated). This is important, as the previous technique did not
allow distinguishing between reads of zero rows, and reads of partial
state.