Rewrite COUNT(*) queries to fit our API
Soup aggregations require an explicit `over` column, which COUNT(*) does
not have. This adds a pass that rewrites parsed queries to pick an
arbitrary column to compute over. This slightly skews COUNT(*)
semantics, as it is supposed to also count rows with NULL values;
however, we don't yet support computed columns on JOIN results anyway,
and our LEFT JOIN is broken, so this will only become an issue when
those are addressed. (The solution is still not ideal: it would be
better to do the rewriting dynamically on graph construction, but having
spent a fair amount of time trying to punch through the required
information, I concluded that the rewrite pass hack is a better plan,
since this feature isn't especially critical.)

With this change, we now faithfully generate a Soup graph for the full
TPC-W query set with the exception of substring queries that aren't
compatible with the Soup model.