Avoid double-replaying updates during replay
The old code had a race where the following sequence of events could
happen:

 - Domain forwards an update u1
 - Domain is told to replay
 - Domain copies state to replay a node
 - Domain sends state
 - Target domain receives instructions to prepare to receive replay
 - Target domain starts buffering updates to target node
 - Target domain receives u1 and buffers it
 - Target domain receives replay state
 - Target domain processes buffered update u1

This results in u1 being double-applied. Once in the replayed state, and
once in the backlog.

This commit fixes this issue by having a domain *not* start buffering
until it receives its first replay message, *and* for a domain that
spawns a replay thread to first send a single, empty replay message to
let the target domain know to start buffering.

All old tests now pass again!