Process recipe additions in the correct order
This fixes a bug related to our use of `HashMap` structures to convey
expressions added to and removed from a recipe compared to another:
because the iteration order of a `HashMap` is undefined, dependent
expressions would sometimes be added before those that they depend on,
even though they were listed in the correct order in the recipe. By
keeping an additional ordered index over the recipe, we make sure that
additions are always processed in the right order.