Last instead of first column in `COUNT(*)` hack
This is a band-aid fix for an unfortunate clash of hacks: to support
`COUNT(*)`, we rewrite it to count over an arbitrary column (previously,
we used the first one). To support aggregation queries without `GROUP
BY` clauses, we also play some tricks with group columns: either we make
one up (a "bogo group column"), or we push a parameter column (which we
must carry all the way to the leaf node for the query anyway) into the
group. However, it is often the case that said made-up group column is
actually the *first* column in the view. This means that a query like

 `SELECT COUNT(*) FROM shopping_cart_line WHERE scl_id = ?;`

ends up breaking, because we use the `scl_id` column *both* as a column
to compute over and as a group column since it is the first column of
`shopping_cart_line`.

This workaround uses the last column; the clash can still happen, but it
doesn't happen in any of our benchmarks and is probably less likely.