Fix off-by one error in transaction fast-forward
With help from @fintelia, found an off-by one error that caused us to
process all supposedly skipped timestamps when a domain fell behind on
transactions (in other words, the fast path never triggered).

Also, we didn't correctly handle subsequent writes to the same base
table; we can speed this up further by not even generating the `HashMap`
with transaction fast-forward information if we know that two writes to
the same base table are on subsequent timestamps.