Reduce ambiguities in implied table expansion
Previously, our implied table expansion considered *all* base table
schemas, whether the table was present in the query or not. This caused
false ambiguities when there were multiple columns of the same name in
different base tables (as in HotCRP): even though a query explicitly
mentioned only one table, the resolution would find both.

In addition to fixing this, refactor the implied table expansion code to
be less confusing and separate the CREATE TABLE and SELECT cases more
clearly.