Consider joined clause for implied table expansion
Columsn in tables mentioned in a `JOIN` clause of a selection can
be referred to from conditionals, so we must consider these tables for
implied table expansion, too. The current version increases ambiguity,
however: the column(s) which are being joined (`USING` or `ON` clause)
necessarily occur in both tables; if they are projected, the current
code gets confused (we need to treat the join column specially here).