Flip join condition if it's the wrong way round
Since users are free to write their join conditions any way they like,
the order in which tables are mentioned in the join condition need not
agree with the order in which the tables are joined. This simple fix
flips a flat join condition around if it is in reverse to the known join
order; it does not handle more complex join conditions with logical
operators.