Track latest schema in which each view appears
It turns out the fix in 17114afea1cdcc82591f2a78ac144e1fd37f004a is not
sufficient to have all nodes (including reused ones) reflected in all
schema versions: sometimes, we already decide to reuse an existing query
at a higher level than the MIR (e.g., at the recipe level or the QG
level). In that case, we never reach the code that adds copies of nodes
reused in the MIR at later schema versions. Instead, we now track the
latest schema version in which nodes of a given name appear, and use
that to retrieve the MIR node for reuse. I believe this makes the changes
in 17114afea1cdcc82591f2a78ac144e1fd37f004a redundant, but (a) they don't
hurt, and (b) explicitly re-registering the nodes if we *do* get to the
MIR level is at least consistent.