Query reuse with different parameters via MIR
If we detect that a query is identical to an existing one in the Soup
graph, except that the query parameters are on different columns, we
reuse the existing query and simply add another leaf node. However, this
was not previously implemented via the MIR: instead, the high-level QG
reuse code manually added flow graph nodes. This change moves that
responsibility into the MIR: such a reuse is now modelled as a (very
short) MIR query that consists of a `Reuse` node, an `Identity` node,
and a `Leaf` node.

This finally fixes the last broken test for MIR-based graph creation!