Fix bug in MIR `Filter` node generation
Previously, the filter node would try to look up its columns in its
parent node's column set to build the set of filters. Sometimes,
however, the parent MIR node did not actually have the necessary column
(due to peephole generation). The column pull-down fixes in 913e6ae1f1
do not address this because they are only applied *after* MIR
generation. Instead, we now simply append unknown columns to the local
filter condition and column vectors and rely on pull-down to actually
make them available when the MIR is optimized (note that this is always
correct, I think, because the columns are appended in the same order
that they will be pulled down in).

We now get to HotCRP v11 again with the MIR-based graph generation
enabled!