Get latest base when making mutator via recipe
There may not be a node at this schema version if nothing about the base
node (or query) has changed, so we need to use our internal bookkeeping
info to find the latest schema version at which it exists.

We now populate HotCRP v11 correctly from the test data set!