Extract join tables from "ON" in explicit join
As noted in ca637beaa5de7ba2158ad4656b238619bf90d9b9, we must actually
determine the left and right tables of each explicit join by analysing
its join predicate. This change introduces code in the QueryGraph
implementation do to so. However, it unfortunately still has to maintain
the hacky logic that flips the ConditionTree in the predicate according
to which column ends up being the left one.