Infrastructure for MIR-level merge for reuse
When we detect an opportunity for query reuse at the QG level, we still
need to translate both the reused query and the new query into MIR, so
that we can then bring in the appropriate `Reuse` nodes in the MIR for
the query being added. This change largely adds infrastructure required
to make such sharing work, but not yet the actual sharing.