Writes and replays convert from transactional to non-transactional
Specifically, both of these packet types drop their timestamps when they pass
from a transactional part of the graph to a non-transactional part. (Or actually
in the case of writes the transformation occurs before the next domain crossing
but not earlier. This could be changed relatively easily, but I haven't taken
the time to do so yet.)