Return column constraints on table creation
This introduces a new type, `ColumnSpecification` that carries not just
the information required to form a `Column` structure, but also a type
and the associated column constraints. It represents a column
description as given in a CREATE TABLE statement, while `Column` is a
more generic description of a table column (e.g., as mentioned in a
SELECT statement).