Tidy up the way Reader state is initialized
The handles inside Readers are now exclusively created by the domain,
which then makes the  ReadHandle globally. This allows us to defer the
materialization decision until commit time, removing the hacky partial
materialization code that swapped the Read and Write handles in the
graph.

The ReadHandle still holds a closure with a Mutex<SyncSender>, but this
is locally constructed, and not sent over a channel. This should
hopefully be more amenable to distribution.