Rewrite leaf view columns to be standalone
This is a small but important change that fixes our ability to refer to
existing views in the definitions of new ones: previously, this was
broken because the MIR leaf nodes would still carry columns that
identified their table as the base table they originated at (or none, in
the case of aggregations). For example, the `aid` column on `VoteCount`
would be stored as `Article.aid`, and view definitions that nest
`VoteCount` had to refer to it as that. This is confusing and also
incorrect, so we now sanitize the column definitions on leaf views and
rewrite them as if they were base table columns. This allows `Leaf`-type
MIR nodes to be used as if they were base tables, and thus allows
arbitrary view composition. Yay!