Lay groundwork for less hacky base node reuse
This makes a start at porting the code from the `hacky-base-reuse`
branch to the current master (albeit rewound prior to the sharding
merge, as the sharding code seems buggy with HotCRP migrations).

The `Base` MIR node now carries an optional `BaseNodeAdaptation` that
specifies columns to add and remove over an earlier base node version.
It also carries the `ColumnSpecification` for each base node column;
this is required in order to convey column types and default values.

This commit ports the logic to detect base node adaptation opportunities
(from `hacky-base-reuse`) and complex schema changes, but does not yet
include the code to action such reuse.