Correctly set filter width on MIR nodes
This generates filter conditions whose width is correct even if the
filter is adjacent to a base node, which may have "hidden" removed
columns (on which we do not wish to filter, but which still need to be
accounted for correctly).

Also includes a fix that sets the width of merged filters correctly:
previously, we would simply use the field count on the first merged
node, but this isn't correct in situations like the one described above.