Weaker assertion in filter node setup
It's okay for the filter to be a prefix of the columns present on the
parent node; we implicitly pad it with `None` entries and ignore the
extra columns. This situation only occurs if the last column (or several
columns at the end) have been deleted and no more columns have been
added; otherwise, we will correctly widen the filter to the maximum
absolute column ID observed.