Version config to allow non-cargo commands
This adds a "version" flag to the Taster config in order to allow
breaking changes to taster's behavior with backwards compatibility to
old configuration files.

The first such change is that from version 2, we no longer assume that
all benchmark commands are implicitly prefixed with the "cargo" command.
This allows running non-Rust commands for benchmarks.