Allow for literal columns in SQL field list
Checkpoints work on projecting literals into the output columns of a
SQL query. This commit adds the necessary data structure indirection to
support both ordinary data columns (a `Column` stored within the
`OutputColumn::Data` enum variant) and literal columns (a `LiteralColum`
stored in `OutputColumn::Literal`). It also stores the output column set
as a separate top-level field on the query graph, rather than relying on
composition of the columns selected from the relations in the QG.

Future changes will replace the leaf projection of a query by one that
includes the literal columns.