Add HTTP API for recipes; prepare Blender linkage
This removes the misguided "recipe" argument I had added, and replaces
it instead with a REST endpoint for POST-ing new recipes. It also
restructures the controller state to hold an `Arc<Mutex<Blender>>`, as
the blender must be shared between the API server (which migrates
between recipes) and the controller (which adds/removes workers).