Unique names for subqueries
Since our anonymous query names are based on the number of queries
currently in the system and we only increment that count *after* a query
has been added, subqueries could end up having the same name as their
containing parent query, which broke a bunch of things.

The fix is to increment the query counter *before* a query and its
subqueries are added.

Same fix as applied by @larat7 as part of
b861fc4d0410a470579121c8ef9e895aade4613d.