Fix race on `UpdateEgress` and channel coordinator
When operating distributedly, actual domain boot and further steps taken
at the controller to connect domains (in the "routing" phase of
migration) can race with each other. Specifically, this means that the
`UpdateEgress` message to the domain (coming from the controller) might
be handled *before* the local channel coordinator has been updated with
the parent domain's whereabouts. Previously, we simply did nothing and
ended up with an egress node without children. Now, we spin until the
channel coordinator has connectivity information for the domain. This
isn't at all fault-tolerant (a domain's machine could crash while it
boots), but fixes the race.