Remove UB in aliased use of Box in State
We previously implemented deduplication using aliased Boxes. Apparently
this is considered undefined behavior in many cases (see relatively long
discussion starting at
https://botbot.me/mozilla/rust/2017-09-07/?msg=90810543&page=18), so we
want to avoid it. This commit changes the internals of State to instead
have records be *mut T with a wrapper type that implements
Deref<Target=T>. This ensures that we never get double-frees (since a
drop of *mut T doesn't do anything), but also means that we need to
explicitly convert the *mut T back into a Box<T> when a drop should
happen.

This change has the added benefit of enabling us to add another index
after the fact that still uses the same de-duplication. To see why,
consider how you would take a &Box<T> and turn it into a Box<T> pointing
to the same memory as the &Box<T>. In this new world, all we need to do
is Row(old_row.0), since *mut T can be copied.