Remove hard-coded IP addresses from channel code
Various places in the channel subcrate used to have hard-coded local
IP-port combinations like "127.0.0.1:0" or "0.0.0.0:0". This change
removes them, and instead changes the constructor APIs for channels and
polling loops to take an explicit SocketAddr. This is required for
distributed scale-up experiments to work without manual patches.

Note that this has just moved the issue to the main codebase now (which
still has hard-coded IPs). These need to be fixed in a next step, but
will change several interfaces, so probably best done in a separate
commit.