ReaderOntoExisting reuse is not always okay
Specifically, if the projected columns of two queries match, this does
*not* mean that we can necessarily simply add another reader to the
existing query: if the query contains any grouped expressions, they are
implicitly grouped by the parameter columns, which we know differ by
definition in this situation. So, even though the projected columns look
identical, they're really different because of the implied GROUP BY
clause.

This fixes the issue in a conservative way: we simply reject any query
that projects a computed column for `ReaderOntoExisting` reuse.

Fixes an error with schema v9 in HotCRP.