Split MIR code into modules
This change reorganizes the MIR implementation from one behemoth file
into separate submodules. This is a clean-up change I've been meaning to
make for a long time, and now seems to be a good time to do it with
minimal breaking of dependent pieces of code or concurrent changes.

Note that this change tries to keep the code as intact as possible,
rather than applying (also sensible) refactorings in the same pass.
Specifically, the current splitting makes a lot more types and methods
public than they really need to be; in the interest of being able to
track code motion, this makes sense here, but I will address it in
future commits.