Move controller thread creation outside build()
This is necessary because we otherwise end up back in the situation
where a `ControllerInner` is sent across thread boundaries (namely,
`self` is sent into the `thread::spawn` closure).
baac511a98c62b61c09640c6ae96d15cde9e7662 laid the ground work for
this in order to support long-term recipe storage in `ControllerInner`,
but this change is required before we can actually store a recipe.