Re-enable local handling
This commit fixes a major issue with local packet handoff.
Specifically, consider what happens if we have a single worker and two
replicas, R1 and R2, where R1 has multiple outputs, all of which lead to
R2. The worker handles a packet to R1, which results in two outputs to
R2, O1 and O2, in that order. The worker decides to locally handle O1,
and put O2 on the TCP channel to R2. Then, another packet arrives for R1
before O2 is handled. The worker picks up that packet, and it produces
only one output: O2'. O2' is for the same node in R2 as O2. Say the
worker now decides to handle O2' locally. Without further checks, this
results in out-of-order operation along a single edge, because O2 should
be processed before O2'.

This commit fixes this problem by having replicas track how many packets
they have sent to/received from other replicas. In the case above, the
worker would see that it has sent two packets to R2, but that only one
has been received (and processed). It therefore aborts the local
handoff, and reverts to sending O2' on the TCP channel instead.

This scheme does have some false positives, since the tracking should
really be per target node rather than per replica, but that would
require some further introspection into packets that I didn't want to
add. We may want to consider adding this in the future if it results in
too many aborts (aborts are somewhat expensive, as they need to take
lock the target replica).