Add workaround for weird epoll race.
Something very weird happens in this particular test, and I don't have a
good explanation for it, nor a good way to fix it beyond what's in this
commit. For posterity:

I'm seeing some truly bizarre behavior where I get events for an fd
after deregistering it. Specifically, I have one thread that registers
the fd with ONESHOT. Strictly later, another thread (correctly) gets an
event for the token for that fd, reads from the fd, and then deregisters
it (+ closes the fd's stream). But then that same thread immediately
gets the same token back from poll()... Unfortunately this doesn't occur
if I run the application through strace (which slows it down), so there
must be a race But I've combed through my code pretty carefully and
can't seem to find a race on the tokens or fds.

I wonder if this is behavior that could be observed due to mio's
internal housekeeping for tokens (it does a bunch of concurrent
datastructure operations to keep an &self interface)? I suspect the
problem is related to token reuse, but have no way of proving this.