Configurable persistence and deployment ID
Souplet now stores its base logs persistently unless --emphemeral is
specified. Also introduces a "deployment ID" that we use for the Zk
entries (as a path prefix) and the log storage (as a filename prefix).
This allows a user to easily re-start Soup for a given application
deployment, and avoids accidentally recovering unrelated state. If no
deployment ID is specified (--deployment), Soup generates a UUID (and
prints it if --verbose is set).

This works with Pan: we can now survive *both* restarts of Pan and of
the Soup deployment, with recipe and base log data preserved and
recovered correctly on restart.