Remove NodeIndex from public mutator/getter API
The RPCs used to obtain read and write handles now no longer take a
`NodeIndex` (which is private information and may change
non-deterministically after recovery), but instead use the name of
the view or base table they target. Impacts tests and benchmarks, but
generally simplifies application code.

Note that it was previously acceptable (sort of) to have multiple base
tables or views with the same name, and to differentiate them by their
`NodeIndex . This is no longer okay, but shouldn't be anyway. Some tests
adapted to fix this.

Note that the `inputs` and `outputs` RPCs continue to exist and still
return `NodeIndex es; we should change them to just return lists of
strings, or remove these RPCs since their primary purpose has been
eliminated.