More correct ExactMatch/ExtendExisting reuse
Previously, column order did not matter in some of our data structures
used for reuse decisions (specifically, in the query graph signature).
This is problematic because it can cause us to incorrectly conclude that
two queries are identical even though they output the columns in a
different order. If that is the case, we should fall into the
`ExtendExisting` reuse case, add a reprojection operator, and add a new
reader. This makes that change and extends the unit test for reuse with
an explicit check to ensure we actually do it.

This likely goes some way to fixing the incorrect and overeager use of
`ExactMatch` observed by @larat7.