Progress towards executing queries
SELECT and INSERT queries now arrive in Soup and appear to work,
although there are a bunch of corner cases not yet dealt with (example:
repeated execution of the same query -- the Soup side needs to correctly
supply a getter for the reused query).

Manages to get past the first SELECT and INSERT queries in Django app
bootup.